适配复读插件
