+++ tests/plugins/repeat/test_history.py
@@ -15,9 +15,12 @@ async def test_history(app: App, mocker: MockerFixture):
     require("src.plugins.repeat")
     from nonebot.adapters.onebot.v11 import Bot, Message
 
-    from src.plugins.repeat import history_cmd, plugin_config, recorder_obj
+    from src.plugins.repeat import plugin_config, recorder_obj
+    from src.plugins.repeat.plugins.history import history_cmd
 
-    mocked_datetime = mocker.patch("src.plugins.repeat.history.datetime")
+    mocked_datetime = mocker.patch(
+        "src.plugins.repeat.plugins.history.data_source.datetime"
+    )
     mocked_datetime.now.return_value = datetime(2020, 1, 2)
     mocked_datetime.return_value = datetime(2020, 1, 1)
 
@@ -55,9 +58,12 @@ async def test_history_get_arg(app: App, mocker: MockerFixture):
     require("src.plugins.repeat")
     from nonebot.adapters.onebot.v11 import Bot, Message
 
-    from src.plugins.repeat import history_cmd, plugin_config, recorder_obj
+    from src.plugins.repeat import plugin_config, recorder_obj
+    from src.plugins.repeat.plugins.history import history_cmd
 
-    mocked_datetime = mocker.patch("src.plugins.repeat.history.datetime")
+    mocked_datetime = mocker.patch(
+        "src.plugins.repeat.plugins.history.data_source.datetime"
+    )
     mocked_datetime.now.return_value = datetime(2020, 1, 2)
     mocked_datetime.return_value = datetime(2020, 1, 1)
 
@@ -106,7 +112,8 @@ async def test_history_get_invalid_args(app: App):
     require("src.plugins.repeat")
     from nonebot.adapters.onebot.v11 import Bot, Message
 
-    from src.plugins.repeat import history_cmd, plugin_config
+    from src.plugins.repeat import plugin_config
+    from src.plugins.repeat.plugins.history import history_cmd
 
     plugin_config.group_id = [10000]
 
@@ -143,7 +150,7 @@ async def test_history_not_enabled(app: App):
     from nonebot.adapters.onebot.v11 import Bot, Message
 
     require("src.plugins.repeat")
-    from src.plugins.repeat import history_cmd
+    from src.plugins.repeat.plugins.history import history_cmd
 
     async with app.test_matcher(history_cmd) as ctx:
         bot = ctx.create_bot(base=Bot)
