+++ tests/plugins/repeat/test_status.py
@@ -8,27 +8,27 @@ from tests.fake import fake_group_message_event
 
 
 @pytest.mark.asyncio
-@pytest.mark.parametrize("app", [("src.plugins.repeat",)], indirect=True)
-async def test_status(
-    app: App,
-    mocker: MockerFixture,
-):
+async def test_status(app: App, mocker: MockerFixture):
     """测试状态"""
-    from nonebot import get_driver
-    from nonebot.adapters.onebot.v11 import Adapter, Bot, Message
-    from nonebug.mixin.call_api.fake import make_fake_adapter, make_fake_bot
+    from nonebot import require
+    from nonebot.adapters.onebot.v11 import Message
 
-    from src.plugins.repeat import recorder_obj, status_cmd
+    require("src.plugins.repeat")
+    from src.plugins.repeat import recorder_obj
+    from src.plugins.repeat.plugins.status import status_cmd
 
     recorder_obj.start_time = datetime(2020, 1, 1, 0, 0, 0)
-    mocked_datetime = mocker.patch("src.plugins.repeat.status.datetime")
+    mocked_datetime = mocker.patch(
+        "src.plugins.repeat.plugins.status.data_source.datetime"
+    )
     mocked_datetime.now.return_value = datetime(2021, 2, 2, 1, 1, 1)
-    mocked_server_status = mocker.patch("src.plugins.repeat.status.server_status")
+    mocked_server_status = mocker.patch(
+        "src.plugins.repeat.plugins.status.data_source.server_status"
+    )
     mocked_server_status.return_value = "test"
 
     async with app.test_matcher(status_cmd) as ctx:
-        adapter = make_fake_adapter(Adapter)(get_driver(), ctx)
-        bot = make_fake_bot(Bot)(adapter, "1")
+        bot = ctx.create_bot()
         event = fake_group_message_event(message=Message("/status"))
 
         ctx.receive_event(bot, event)
